+++ bydbctl/internal/cmd/health_check_test.go
@@ -32,7 +32,7 @@ import (
 
 var _ = Describe("health check after launching banyandb server", func() {
 	var deferFunc func()
-	var grpcAddr, certFile string
+	var grpcAddr, httpAddr, certFile string
 	var rootCmd *cobra.Command
 	BeforeEach(func() {
 		_, basePath, _, _ := runtime.Caller(0)
@@ -41,12 +41,21 @@ var _ = Describe("health check after launching banyandb server", func() {
 		}
 		certFile = filepath.Join(basePath, "test/integration/standalone/other/testdata/server_cert.pem")
 		keyFile := filepath.Join(basePath, "test/integration/standalone/other/testdata/server_key.pem")
-		grpcAddr, _, deferFunc = setup.StandaloneWithTLS(certFile, keyFile)
+		grpcAddr, httpAddr, deferFunc = setup.StandaloneWithTLS(certFile, keyFile)
 		rootCmd = &cobra.Command{Use: "root"}
 		cmd.RootCmdFlags(rootCmd)
 	})
 
-	It("should pass", func() {
+	It("http health check should pass", func() {
+		rootCmd.SetArgs([]string{"health", "--addr", "http://" + httpAddr})
+		out := capturer.CaptureStdout(func() {
+			err := rootCmd.Execute()
+			Expect(err).NotTo(HaveOccurred())
+		})
+		Expect(out).To(ContainSubstring("SERVING"))
+	})
+
+	It("grpc health check should pass", func() {
 		rootCmd.SetArgs([]string{"health", "--grpc-addr", grpcAddr, "--grpc-cert", certFile, "--enable-tls", "true"})
 		out := capturer.CaptureStdout(func() {
 			err := rootCmd.Execute()
@@ -55,7 +64,7 @@ var _ = Describe("health check after launching banyandb server", func() {
 		Expect(out).To(ContainSubstring("connected"))
 	})
 
-	It("should pass with insecure flag set", func() {
+	It("grpc health check should pass with insecure flag set", func() {
 		rootCmd.SetArgs([]string{"health", "--grpc-addr", grpcAddr, "--insecure", "true", "--enable-tls", "true"})
 		out := capturer.CaptureStdout(func() {
 			err := rootCmd.Execute()
@@ -64,7 +73,7 @@ var _ = Describe("health check after launching banyandb server", func() {
 		Expect(out).To(ContainSubstring("connected"))
 	})
 
-	It("should fail without the proper cert", func() {
+	It("grpc health check should fail without the proper cert", func() {
 		rootCmd.SetArgs([]string{"health", "--grpc-addr", grpcAddr, "--enable-tls", "true"})
 		err := rootCmd.Execute()
 		Expect(err).To(HaveOccurred())
@@ -82,7 +91,7 @@ var _ = Describe("health check without launching banyandb server", func() {
 		cmd.RootCmdFlags(rootCmd)
 	})
 
-	It("should fail", func() {
+	It("http health check should fail", func() {
 		rootCmd.SetArgs([]string{"health"})
 		err := rootCmd.Execute()
 		Expect(err).To(HaveOccurred())