+++ CHANGES.md
@@ -21,6 +21,7 @@ Release Notes.
 - Bump node to 20.9.
 - Bump several tools.
 - Bump all dependencies of Go and Node.
+- Combine banyand and bydbctl Dockerfile.
 
 ## 0.5.0